Add email cancellation feature and update related services
- Implemented `SendEmail` method in `EmailService` to send emails with a body parameter.
- Updated `IEmailService` interface to include the new method signature.
- Added `CancelAppointmentTemplate` to `MailTemplates` enum.
- Enhanced `AppointmentManager` with `UpdateAppointment` and `GetAppointmentById` methods.
- Updated `IAppointmentManager` interface to reflect new methods.
- Integrated `CancelationService` in `DoctorController` for handling reservation cancellations.
- Modified `AddReview` action to validate user roles and appointment status.
- Adjusted `PatientController` to handle nullable reservation IDs.
- Updated `PaymentController` to cast `DoctorReservationId` to integer.
- Cleaned up `MVC.csproj` by removing references to `Services` folder and unused TypeScript files.
- Added review mapping method in `DoctorMapper` and `IDoctorMapper`.
- Registered `CancelationService` in `Program.cs` for dependency injection.
- Updated `AddReviewVM` to include a required `Rate` property.
- Changed delete behavior for `DoctorReservationID` in `AppDbContext` to `SetNull`.
- Enhanced `AppointmentRepository` with a method for conditional retrieval of appointments.
- Created `CancelationService` class to manage reservation cancellations and send emails.
- Defined `ICancelationService` interface for cancellation functionality.

+++ MVC/Controllers/PatientController.cs
@@ -66,7 +66,7 @@ namespace MVC.Controllers
         }
         public async Task<IActionResult> CancelAppointment(int appointmentId)
         {
-            orderManager.DeleteOrder(int.Parse(User.FindFirst("currentId").Value), await appointmentManager.GetReservationId(appointmentId));
+            orderManager.DeleteOrder(int.Parse(User.FindFirst("currentId").Value),(int) await appointmentManager.GetReservationId(appointmentId));
             await appointmentManager.DeleteAppointmentAsync(appointmentId);
             return RedirectToAction("Profile");
         }
